removed old login page
